Move SEP code from framework IME to framework App
